Move components into `features/models/workspace`
