+++ web-common/src/features/metrics/workspace/MetricsEntityTable.svelte
