+++ web-common/src/features/metrics/workspace/MetricsGenerateButton.svelte
@@ -6,7 +6,7 @@
   import {
     addQuickMetricsToDashboardYAML,
     MetricsInternalRepresentation,
-  } from "./metrics-internal-store";
+  } from "../metrics-internal-store";
   import QuickMetricsModal from "./QuickMetricsModal.svelte";
 
   export let selectedModel: V1Model;
