+++ web-common/src/features/metrics/workspace/MetricsLayoutManager.svelte
