+++ web-common/src/features/metrics/workspace/QuickMetricsModal.svelte
